dont report negative core memory usage
git-svn-id: https://svn.atnf.csiro.au/difx@7244 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ mpifxcorr/trunk/src/core.cpp
@@ -136,7 +136,7 @@ Core::Core(int id, Configuration * conf, int * dids, MPI_Comm rcomm)
   processthreads = new pthread_t[numprocessthreads];
   processconds = new pthread_cond_t[numprocessthreads];
   processthreadinitialised = new bool[numprocessthreads];
-  threadbytes = new int[numprocessthreads];
+  threadbytes = new long long[numprocessthreads];
   for(int i=0;i<numprocessthreads;i++)
   {
     pthread_cond_init(&processconds[i], NULL);
